Crate seg_tree

A generic implimentation of a segment tree (or range query tree), relying on a simple Merge trait to define the operation used during range queries. The zero-recursion approach is taken from this excellent GeeksForGeeks article.

While powerful, the trait-based appoach does have one major disadvantage: If are using this crate as an external dependency, and you want to impliment Merge on a type defined outside the crate in which you are working (e.g. i32), you will have to wrap it in a “dummy” type defined inside your crate. This is because Rust will not allow you to impliment an external trait on an external type. There is no runtime overhead for this wrapping, but it can make your code a little verbose.

Example of wrapping an i32 to impliment Merge on it:

use seg_tree::{SegTree, Merge};
 
#[derive(Default, Clone, Debug, PartialEq)]
struct MyInt(i32);
 
impl Merge for MyInt {
   fn merge(l: &Self, r: &Self) -> Self { MyInt(l.0 + r.0) }
}
 
fn main() {
    let mut tree: SegTree<_> = vec![1,2,-1,4].into_iter().map(|n| MyInt(n)).collect();
    assert_eq!(tree.query(0, 4), MyInt(6));
    assert_eq!(tree.query(1, 3), MyInt(1));
    tree.update(2, MyInt(3));
    assert_eq!(tree.query(0, 4), MyInt(10));
    assert_eq!(tree.query(1, 3), MyInt(5));
}

Alternatively, you could simply copy-paste the source into your own crate and impliment Merge on whatever type you like!

Structs

SegTree

A data structure for allowing both range queries and point updates on a sequence in O(log(n)) time.

Traits

Merge

A trait for defining the operation that should be used when accumulating values in a range query. Although not required by the type system, this operation should be associative for the results of SegTree::query to make any sense.