Trait seg_tree::Merge

pub trait Merge {
    fn merge(l: &Self, r: &Self) -> Self;
}

A trait for defining the operation that should be used when accumulating values in a range query. Although not required by the type system, this operation should be associative for the results of SegTree::query to make any sense.

Required Methods

fn merge(l: &Self, r: &Self) -> Self